API Reference

IFTDuals exports the following functions

Main function

IFTDuals.ift — Function

    function ift(y::Union{V,<:AbstractArray{V}},f::Function,args; DT::Type{Union{Nothing,<:Dual}}=nothing, tag_is_mixed::Union{Nothing,Bool}=nothing, args_needs_promotion::Bool=true) where {V<:Real}
    function ift(y::Union{V,<:AbstractArray{V}},f::Function,args,args_primal; DT::Type{Union{Nothing,<:Dual}}=nothing, tag_is_mixed::Union{Nothing,Bool}=nothing, args_needs_promotion::Bool=true) where {V<:Real}

Function to compute higher-order derivatives using the implicit function theorem (IFT) and (nested) Dual numbers. Input:

y : primal input solution to the root finding problem (scalar or vector)

f : function handle that takes 'y' and 'args' as inputs. f(y,args) = 0 is assumed to define the implicit relationship between y and values given as args.

args : tuple or data structure containing Dual numbers indicating the differentiation structure

args_primal : primal values of args

Optional Input:

DT : Target Dual type for the output. If not provided, it is inferred from args.

tag_is_mixed : Boolean indicating if mixed tags are present in args. If not provided, it is inferred from args.

args_needs_promotion: Boolean indicating if args need to be promoted to a common Dual type.

Output:

Returns y as a Dual number with the appropriate order and partial derivatives computed using the IFT.

The function works by first determining the order of differentiation and whether mixed tags are present.

Warning

When mixed tags are detected, but tag_is_mixed=false is provided, a check is performed to ensure that the number of partials are consistent across the Dual types in args. If inconsistencies are found, an error is thrown. No other checks are performed and it is the user's responsibility to ensure symmetry of partial derivatives in this case.

Warning

If args_needs_promotion=false is provided, it is the user's responsibility to ensure that all Dual types in args are of the same type. If the Dual type differs, unexpected behavior may occur.

The ift function is the core function for computing higher-order derivatives. Using this function will work for most use cases without needing to overload the utility functions. It is a requirement that all functions called on y and args use AbstractArray annotations and not concrete Vector, Matrix or Array types. Internally ift calls some functions which convert arrays of duals into AbstractArray types, to avoid allocating new arrays and hence improve performance.

The ift function has several optional keyword arguments to optimize performance in specific scenarios. These are described in detail in the Advanced Usage section of the user guide.

Utility functions

IFTDuals.pvalue — Function

    pvalue(x::V) where V
    pvalue(::Type{V}) where V

Extracts the value field of Duals contained within generic data structures. pvalue(x::V) return ForwardDiff.value(x), and pvalue(::Type{V}) returns ForwardDiff.valtype(V). For structures containing multiple types (e.g. Tuples, Structs, Arrays of structs, Dicts) it loops through all fields/entries in the data structure and performs pvalue on each entry. For custom structs it is recommended to provide your own method for pvalue which only applies pvalue to the fields which may contain Duals.

Example:

struct MyStruct{T1<:Real,T2<:Real}
    a::T1 # variable which may contain Duals
    b::String # non-numeric variable
    c::T2 # variable which may contain Duals
end

pvalue(x::MyStruct{T1,T2}) where {T1<:Real,T2<:Real} = MyStruct{pvalue(T1),pvalue(T2)}(pvalue(x.a), x.b, pvalue(x.c)) # construct new struct with primal eltypes, need not loop through all fields

We encourage the usage of pvalue to ensure x.a and/or x.c are correctly handled if they are custom structs themselves.

IFTDuals.nested_pvalue — Function

    nested_pvalue(::Type{V}) where V
    nested_pvalue(x::V) where V

Extracts the innermost primal value or types from generic structures containing Duals. It recursively applies pvalue until the innermost primal value/type is not a Dual. Similar to pvalue, for custom structs nested_pvalue is aplied to all fields in the struct. For custom structs it is recommended to provide your own method for nested_pvalue, similar to pvalue.

Example:

struct MyStruct{T1<:Real,T2<:Real}
    a::T1 # variable which may contain Duals
    b::String # non-numeric variable
    c::T2 # variable which may contain Duals
end

nested_pvalue(x::MyStruct{T1,T2}) where {T1<:Real,T2<:Real} = MyStruct{nested_pvalue(T1),nested_pvalue(T2)}(nested_pvalue(x.a), x.b, nested_pvalue(x.c))

We encourage the usage of nested_pvalue to ensure x.a and/or x.c are correctly handled if they are custom structs themselves.

IFTDuals.promote_common_dual_type — Function

    promote_common_dual_type(x::V, DT::Type{<:Dual}) where V
    promote_common_dual_type(::Type{V1}, ::Type{V2}) where {V1,V2}

Promotes Dual numbers in generic data structures to a common Dual type DT by using Base.convert. Or when called on types, it checks if V1 can be promoted to V2 and returns V2 if possible. For structs, the current implementation applies promote_common_dual_type to all fields in the struct and attempts to reconstruct the struct using the promoted values. It is hence recommended to provide your own method for custom structs which only promotes the relevant fields.

Example:

struct MyStruct{T1<:Real,T2<:Real}
    a::T1 # variable which may contain Duals
    b::String # non-numeric variable
    c::T2 # variable which may contain Duals
end

promote_common_dual_type_structs(x::MyStruct{T1,T2},DT::Type{<:Dual}) where {T1<:Real,T2<:Real} = begin 
    TT1 = promote_common_dual_type(T1, DT) # should be DT, but internally checks that this conversion is possible
    TT2 = promote_common_dual_type(T2, DT) # should be DT, but internally checks that this conversion is possible
    MyStruct{TT1,TT2}(promote_common_dual_type(x.a,DT), x.b, promote_common_dual_type(x.c,DT))
end
promote_common_dual_type(x::MyStruct{T,T},DT::Type{T}) where T = x # already desired type

We encourage the usage of promote_common_dual_type to ensure x.a and/or x.c are correctly handled if they are custom structs themselves.

IFTDuals.promote_my_type — Function

    promote_my_type(::Type{T}) where T
    promote_my_type(x::T) where T

Get the common numeric supertype (Duals) from generic data structures. For non-numeric types (String, Symbol, Nothing, Missing, Function), it returns Nothing. For custom data structures the current implementation checks all (nested) fields for numeric types, extracts these numeric types and reduces Base.promote_type over all numeric types found. It is hence highly recommended to provide your own method for custom data structures.

Example:

struct MyStruct{T1<:Real,T2<:Real}
    a::T1 # variable which may contain Duals
    b::String # non-numeric variable
    c::T2 # variable which may contain Duals
end

promote_my_type(::MyStruct{T1,T2}) where {T1<:Real,T2<:Real} = promote_my_type(T1,T2)
promote_my_type(::Type{MyStruct{T1,T2}}) where {T1<:Real,T2<:Real} = promote_my_type(T1,T2) # needed if promote_my_type is called on Type

The above implementation uses promote_my_type, but in this specific case one could use promote_type directly:

promote_my_type(::MyStruct{T1,T2}) where {T1<:Real,T2<:Real} = promote_type(T1,T2)
promote_my_type(::Type{MyStruct{T1,T2}}) where {T1<:Real,T2<:Real} = promote_type(T1,T2)

We encourage the usage of promote_my_type to ensure T1 and/or T2 are correctly handled if they are custom structs themselves. If the struct signature does not contain type information, we perform the promotion on the fields instead:

promote_my_type(x::MyStruct) = promote_my_type(x.a, x.c)

If MyStruct contained a single numeric field (which may contain duals), MyStruct{T} where T<:Real, we would return promote_my_type(T).

Note

This is hence a combination of Base.eltype and Base.promote_type but specialized to only consider numeric types

Implicit function signature requirements

The ift function only supports the two-argument function signature f(x, args). If your implicit function requires multiple separate parameter arguments, you should define args as a tuple and use splatting within your function:

# If you have a function with multiple parameters
function my_function(x, α, β, γ)
    return ... # some operation
end

# Wrap it to use the f(y, theta) signature
f(x, args) = my_function(x, args...)

# Define theta as a tuple of parameters
args = (α, β, γ)

# Now you can use ift
x_dual = ift(x, f, args)