API Reference

IFTDuals.NumEltype

IFTDuals.check_eltypes

IFTDuals.make_dual

IFTDuals.custom_stack

IFTDuals.extract_partials_

IFTDuals.extract_values_field_from_partials

IFTDuals.get_common_dual_type

IFTDuals.ift

IFTDuals.ift_

IFTDuals.ift_recursive

IFTDuals.nested_pvalue

IFTDuals.nested_pvalue_type

IFTDuals.promote_common_dual_type

IFTDuals.seed_nested_dual

IFTDuals.pvalue

IFTDuals.solve_ift

IFTDuals.NumEltype — Method

    NumEltype(x::V) where V
    NumEltype(::Type{V}) where V

Extracts the numeric eltype(s) from generic data structures. For structures containing multiple types (e.g. Tuples, Structs, Arrays of structs, Dicts) it returns the eltype of each entry as a Tuple or Array. If the eltype is not a number, we return Nothing. This function is used to check the Dual types contained within data structures. For efficiency, you may provide your own methods for custom data structures e.g.

struct MyStruct{T<:Number}
    a::T # variable which may contain Duals
    b::String # non-numeric variable
end

NumEltype(x::MyStruct{T}) where T<:Number = T
NumEltype(::Type{MyStruct{T}}) where T<:Number = T

Important: If the eltypes is Real or Any, we require x and not just the type V. This is because Real and Any are abstract types used for common propomotion, and do not provide concrete information about the actual types contained within the data structure. If you attempt to pass NumEltype(::Type{Real}) or NumEltype(::Type{Any}), a RealTypeError or AnyTypeError will be thrown, respectively. Instead, provide an instance of the data structure to extract the concrete eltypes.

IFTDuals.check_eltypes — Method

    check_eltypes(::Type{V}) where V
    check_eltypes(x::Tuple)

Checks if the eltypes contain any Duals. If it does, returns true. This is a helper function intended for internal use.

IFTDuals.make_dual — Method

    make_dual(y::V,DT::Type{<:Dual{T,V,N}},PT::Type{<:Partials{N,V}},parts::Union{V,<:AbstractVector{V}}) where {T,V,N}

Creates a Dual number of type DT with value y and partial derivatives given by parts, which can be a scalar or a vector of partial derivatives. If y is a vector, it creates a Vector of Dual numbers accordingly.

IFTDuals.custom_stack — Method

    custom_stack(x::Union{V,AbstractArray{V}}) where {V<:Dual}
    custom_stack(x::Union{<:Dual{T,V,N},AbstractArray{<:Dual{T,V,N}}},idx::Union{T2,Vector{T2}}) where {T,V,N,T2<:Union{Int,CartesianIndex{1}}}

Stacks the partial derivatives of a Dual (or Vector of Duals) into a Matrix of M x N, where M is the size of the input, N is the number of partials. When an index idx is provided, it returns a Vector, the 'idx'-th partial derivative(s) (idx-th column of the stacked Matrix but efficiently).

IFTDuals.extract_partials_ — Method

    extract_partials_(x::Dual{T,V,N}) where {T,N,V}
    extract_partials_(x::Dual{T,V,N},idx::Union{T2,Vector{T2}}) where {T,V,N,T2<:Union{Int,CartesianIndex{1}}}

Extracts the partials field as a Tuple from a ForwardDiff.Dual. When an index idx is provided, it extracts the idx-th partial derivative.

IFTDuals.extract_values_field_from_partials — Method

    extract_values_field_from_partials(parts::Partials{N,V}) where {N,V}

Extracts the Tuple from a ForwardDiff.Partials struct. If N==1, it returns a scalar instead of a one-element Tuple.

IFTDuals.get_common_dual_type — Method

    get_common_dual_type(x)

Gets the common Dual supertype from the Dual numbers contained within generic data structures. If now Duals present, returns the common numeric supertype. And if no numeric types are present, it should error.

IFTDuals.ift — Method

    ift(y::Union{V,<:AbstractVector{V}},f::Function,tups) where {V<:Real}

Function to compute higher-order derivatives using the implicit function theorem and (nested) Dual numbers. Input: y : primal input solution to the root finnding problem (scalar or vector) f : function handle that takes 'y' and 'tups' as inputs tups : tuple or data structure containing Dual numbers indicating the differentiation structure

f(y,tups) = 0 is assumed to define the implicit relationship between y and values given as tups.

Note: This function currently does not support mixed-mode AD, i.e. differentiating wrt different variables given as nested Duals. As a workaround you may concatenate all variables into a single vector and differentiate jointly.

IFTDuals.ift_ — Method

    ift_(y::Union{V,<:AbstractVector{V}},BNi::Union{Dual{T,V,N},<:AbstractVector{Dual{T,V,N}}},neg_A::Union{V2,<:LU{V2,<:AbstractMatrix{V2},<:AbstractVector{<:Integer}}}) where {T,V<:Real,N,V2<:Real}

For a given order of differentiation, recusrively computes all directional derivatives using the implicit function theorem and recreates the appropriate ForwardDiff.Dual structure.

IFTDuals.ift_recursive — Method

    ift_recursive(y::Union{V,<:AbstractVector{V}},f::Function,tups,neg_A::Union{V2,<:LU{V2,<:AbstractMatrix{V2},<:AbstractVector{<:Integer}}},der_order::Int) where {V<:Real,V2<:Real}

Recursively applies the implicit function theorem to compute higher-order derivatives up to der_order. It evaluates the function f at the current y, solves for the directional derivatives using ift_, and promotes y to the next order of Dual numbers as needed.

IFTDuals.nested_pvalue — Method

    nested_pvalue(x::V) where V

Extracts the innermost primal value type(s) from generic data structures containing nested Dual numbers. For V<:Dual, it recursively extracts the valtype until reaching a non-Dual type. For structures containing multiple types (e.g. Tuples, Structs, Arrays of structs, Dicts) it returns the innermost primal value types of each entry as a Tuple or Array. For non-Dual types, it returns the input type as is.

For generic structure, we have an implementation which attempts to reconstruct the struct type with primal eltypes, but is possible that this may fail. You may consider providing your own method for custom data structures, using recursive calls to pvalue or ForwardDiff.value as needed.

IFTDuals.nested_pvalue_type — Method

    nested_pvalue_type(::Type{V}) where V

Extracts the innermost primal value type(s) from generic data structures containing nested Dual numbers. For V<:Dual, it recursively extracts the ForwardDiff.valtype(V) until reaching a non-Dual. For structures containing multiple types (e.g. Tuples, Structs, Arrays of structs, Dicts) it returns the innermost primal value types of each entry as a Tuple or Array. For non-Dual types, it returns the input type as is.

This function is used to reconstruct struct types with primal eltypes, maintaining eltypes in struct definitions. For generic structure, we have an implementation which attempts to reconstruct the struct type with primal eltypes, but is possible that this may fail. See nested_pvalue for more details.

IFTDuals.promote_common_dual_type — Method

    promote_common_dual_type(x::V, DT::Type{<:Dual}) where V

Promotes Dual numbers in generic data structures to a common Dual type DT. For V<:Dual, it constructs a new instance of DT with the same value and partials as x.

IFTDuals.seed_nested_dual — Method

    seed_nested_dual(y::V,::Type{<:Dual{T,V,N}}) where {T,V,N})

Promotes a Dual number to one of order + 1, seeding the new directional derivatives with zeros. Or if the input is of concrete type Int,Float64,etc, it simply constructs a Dual number of the target type with zero partials.

IFTDuals.pvalue — Method

    pvalue(x::V) where V

Extracts the primal value(s) from generic data structures containing Dual number. For V<:Number, it returns ForwardDiff.value(x) For structures containing multiple types (e.g. Tuples, Structs, Arrays of structs, Dicts) it returns the primal values of each entry as a Tuple or Array. For non-Dual types, it returns the input as is.

It defaults to returning the input as is. You may consider the following example:

struct MyStruct{T<:Number}
    a::T # variable which may contain Duals
    b::String # non-numeric variable
end

pvalue(x::MyStruct{T}) where T<:Number = MyStruct{ForwardDiff.valtype(T)}(pvalue(x.a), x.b) # construct new struct with primal eltypes

IFTDuals.solve_ift — Method

    solve_ift(y::Union{V,<:AbstractVector{V}},BNi::Union{V,AbstractVecOrMat{V}},neg_A::Union{V,<:LU{V,<:AbstractMatrix{V},<:AbstractVector{<:Integer}}},DT::Type{<:Dual{T,V,N}}) where {T,V<:Real,N}

Solves the implicit function theorem system for a single directional derivative and constructs Dual numbers of type DT with the computed partial derivatives. This is part of a recursive process to compute higher-order derivatives.